requirements.txt
# Core
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.4
python-multipart>=0.0.9

# Scheduling
apscheduler>=3.10.4

# Scraping
python-jobspy>=1.1.79

# Indexer / RAG
gitpython>=3.1.43
chromadb>=0.5.0
sentence-transformers>=3.0.0

# HTTP
requests>=2.32.0

# Data
pandas>=2.2.0
